+++ src/suou/collections.py
@@ -28,7 +28,7 @@ class TimedDict(dict[_KT, _VT]):
     """
     Dictionary where keys expire after the defined time to live, expressed in seconds.
 
-    NEW 0.5.0
+    *New in 0.5.0*
     """
     _expires: dict[_KT, int]
     _ttl: int