+++ src/suou/legal.py
@@ -18,6 +18,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 
 # TODO more snippets
 
+from .strtools import SpitText
+
+
 INDEMNIFY = """
 You agree to indemnify and hold harmless {0} from any and all claims, damages, liabilities, costs and expenses, including reasonable and unreasonable counsel and attorney’s fees, arising out of any breach of this agreement.
 """
@@ -27,7 +30,7 @@ Except as represented in this agreement, the {0} is provided “AS IS”. Other 
 """
 
 GOVERNING_LAW = """
-These terms of services are governed by, and shall be interpreted in accordance with, the laws of {0}. You consent to the sole jurisdiction of {1} for all disputes between You and , and You consent to the sole application of {2} law for all such disputes.
+These terms of services are governed by, and shall be interpreted in accordance with, the laws of {0}. You consent to the sole jurisdiction of {1} for all disputes between You and {2}, and You consent to the sole application of {3} law for all such disputes.
 """
 
 ENGLISH_FIRST = """
@@ -45,5 +48,51 @@ If one clause of these Terms of Service or any policy incorporated here by refer
 """
 
 COMPLETENESS = """
-These Terms, together with the other policies incorporated into them by reference, contain all the terms and conditions agreed upon by You and {{ app_name }} regarding Your use of the {{ app_name }} service. No other agreement, oral or otherwise, will be deemed to exist or to bind either of the parties to this Agreement.
+These Terms, together with the other policies incorporated into them by reference, contain all the terms and conditions agreed upon by You and {0} regarding Your use of the {0} service. No other agreement, oral or otherwise, will be deemed to exist or to bind either of the parties to this Agreement.
 """
+
+
+class Lawyer(SpitText):
+    """
+    A tool to ease the writing of Terms of Service for web apps.
+
+    NOT A REPLACEMENT FOR A REAL LAWYER AND NOT LEGAL ADVICE
+
+    *New in 0.11.0*
+    """
+
+    def __init__(self, /,
+        app_name: str,      domain_name: str,
+        company_name: str,  jurisdiction: str,
+        country: str,       country_adjective: str
+    ):
+        self.app_name = app_name
+        self.domain_name = domain_name
+        self.company_name = company_name
+        self.jurisdiction = jurisdiction
+        self.country = country
+        self.country_adjective = country_adjective
+
+    def indemnify(self):
+        return self.format(INDEMNIFY, 'app_name')
+
+    def no_warranty(self):
+        return self.format(NO_WARRANTY, 'app_name', 'company_name')
+
+    def governing_law(self) -> str:
+        return self.format(GOVERNING_LAW, 'country', 'jurisdiction', 'app_name', 'country_adjective')
+
+    def english_first(self) -> str:
+        return ENGLISH_FIRST
+    
+    def expect_updates(self) -> str:
+        return self.format(EXPECT_UPDATES, 'app_name')
+
+    def severability(self) -> str:
+        return SEVERABILITY
+
+    def completeness(self) -> str:
+        return self.format(COMPLETENESS, 'app_name')
+
+# This module is experimental and therefore not re-exported into __init__
+__all__ = ('Lawyer',)